Fix snmp_index overflow: Integer→BigInteger for Huawei OLT ifIndex values
Huawei OLT SNMP indexes encode slot/port/ONT into the ifIndex value,
producing values like 4194316032 that exceed PostgreSQL Integer range.
Changed to BigInteger to accommodate all vendor index formats.

+++ alembic/versions/h1b2c3d4e5f6_add_snmp_counter_columns_to_interfaces.py
@@ -25,7 +25,7 @@ def upgrade() -> None:
     columns = [c["name"] for c in inspector.get_columns("device_interfaces")]
 
     if "snmp_index" not in columns:
-        op.add_column("device_interfaces", sa.Column("snmp_index", sa.Integer(), nullable=True))
+        op.add_column("device_interfaces", sa.Column("snmp_index", sa.BigInteger(), nullable=True))
     if "last_in_octets" not in columns:
         op.add_column("device_interfaces", sa.Column("last_in_octets", sa.Float(), nullable=True))
     if "last_out_octets" not in columns:
